+++ azext_iot/central/services/device_template.py
@@ -3,7 +3,7 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
-# This is largely derived from https://docs.microsoft.com/en-us/rest/api/iotcentral/devicetemplates
+# This is largely derived from https://learn.microsoft.com/en-us/rest/api/iotcentral/devicetemplates
 
 import requests
 from typing import List
@@ -139,7 +139,7 @@ def create_device_template(
         payload: see example payload available in
             <repo-root>/azext_iot/tests/central/json/device_template_int_test.json
             or check here for more information
-            https://docs.microsoft.com/en-us/rest/api/iotcentral/devicetemplates
+            https://learn.microsoft.com/en-us/rest/api/iotcentral/devicetemplates
         token: (OPTIONAL) authorization token to fetch device details from IoTC.
             MUST INCLUDE type (e.g. 'SharedAccessToken ...', 'Bearer ...')
         central_dns_suffix: {centralDnsSuffixInPath} as found in docs
@@ -182,7 +182,7 @@ def update_device_template(
         payload: see example payload available in
             <repo-root>/azext_iot/tests/central/json/device_template_int_test.json
             or check here for more information
-            https://docs.microsoft.com/en-us/rest/api/iotcentral/devicetemplates
+            https://learn.microsoft.com/en-us/rest/api/iotcentral/devicetemplates
         token: (OPTIONAL) authorization token to fetch device details from IoTC.
             MUST INCLUDE type (e.g. 'SharedAccessToken ...', 'Bearer ...')
         central_dns_suffix: {centralDnsSuffixInPath} as found in docs