+++ azext_iot/central/services/job.py
@@ -3,7 +3,7 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
-# This is largely derived from https://docs.microsoft.com/en-us/rest/api/iotcentral/jobs
+# This is largely derived from https://learn.microsoft.com/en-us/rest/api/iotcentral/jobs
 
 import requests
 from typing import List, Union
@@ -359,7 +359,7 @@ def create_job(
         content: see example payload available in
             <repo-root>/azext_iot/tests/central/json/job_int_test.json
             or check here for more information
-            https://docs.microsoft.com/en-us/rest/api/iotcentral/1.1-preview/jobs/create#cloudpropertyjobdata
+            https://learn.microsoft.com/en-us/rest/api/iotcentral/1.1-preview/jobs/create#cloudpropertyjobdata
         job_name: (OPTIONAL)(non-unique) human readable name for the job
         description: (OPTIONAL) Detailed description of the job.
         token: (OPTIONAL) authorization token to fetch job details from IoTC.