+++ azext_iot/sdk/iothub/service/operations/devices_operations.py
@@ -42,9 +42,9 @@ class DevicesOperations(object):
         """Gets the identities of multiple devices from the IoT Hub identity
         registry. Not recommended. Use the IoT Hub query API to retrieve device
         twin and device identity information. See
-        https://docs.microsoft.com/en-us/rest/api/iothub/service/queryiothub
+        https://learn.microsoft.com/en-us/rest/api/iothub/service/queryiothub
         and
-        https://docs.microsoft.com/en-us/azure/iot-hub/iot-hub-devguide-query-language
+        https://learn.microsoft.com/en-us/azure/iot-hub/iot-hub-devguide-query-language
         for more information.
 
         :param top: The maximum number of device identities returned by the
@@ -290,7 +290,7 @@ class DevicesOperations(object):
     def get_twin(
             self, id, custom_headers=None, raw=False, **operation_config):
         """Gets the device twin. See
-        https://docs.microsoft.com/azure/iot-hub/iot-hub-devguide-device-twins
+        https://learn.microsoft.com/azure/iot-hub/iot-hub-devguide-device-twins
         for more information.
 
         :param id: The unique identifier of the device.
@@ -349,7 +349,7 @@ class DevicesOperations(object):
     def replace_twin(
             self, id, device_twin_info, if_match=None, custom_headers=None, raw=False, **operation_config):
         """Replaces the tags and desired properties of a device twin. See
-        https://docs.microsoft.com/azure/iot-hub/iot-hub-devguide-device-twins
+        https://learn.microsoft.com/azure/iot-hub/iot-hub-devguide-device-twins
         for more information.
 
         :param id: The unique identifier of the device.
@@ -422,7 +422,7 @@ class DevicesOperations(object):
     def update_twin(
             self, id, device_twin_info, if_match=None, custom_headers=None, raw=False, **operation_config):
         """Updates the tags and desired properties of a device twin. See
-        https://docs.microsoft.com/azure/iot-hub/iot-hub-devguide-device-twins
+        https://learn.microsoft.com/azure/iot-hub/iot-hub-devguide-device-twins
         for more information.
 
         :param id: The unique identifier of the device.
@@ -494,7 +494,7 @@ class DevicesOperations(object):
     def invoke_method(
             self, device_id, direct_method_request, custom_headers=None, raw=False, **operation_config):
         """Invokes a direct method on a device. See
-        https://docs.microsoft.com/azure/iot-hub/iot-hub-devguide-direct-methods
+        https://learn.microsoft.com/azure/iot-hub/iot-hub-devguide-direct-methods
         for more information.
 
         :param device_id: The unique identifier of the device.