+++ azext_iot/iothub/params.py
@@ -5,7 +5,6 @@
 # --------------------------------------------------------------------------------------------
 
 from azext_iot.iothub.providers.state import HubAspects
-from azext_iot.iothub.common import CertificateAuthorityVersions
 from azure.cli.core.commands.parameters import get_enum_type, get_three_state_flag
 from azext_iot.common.shared import DeviceAuthType, SettleType, ProtocolType, AckType
 from azext_iot.assets.user_messages import info_param_properties_device
@@ -657,12 +656,4 @@ def load_iothub_arguments(self, _):
             "system_properties",
             options_list=["--system-properties", "--sp"],
             help="System properties of the route message.",
-        )
-
-    with self.argument_context("iot hub certificate root-authority set") as context:
-        context.argument(
-            "ca_version",
-            options_list=["--certificate-authority", "--cav"],
-            help="Certificate Root Authority version. The v1 represents Baltimore CA and v2 represents Digicert CA.",
-            arg_type=get_enum_type(CertificateAuthorityVersions),
         )