+++ HISTORY.rst
@@ -512,7 +512,7 @@ Other notable changes, which are not affected by API versions, include:
 **Device Update**
 
 * Introducing the **in preview** Azure Device Update for IoT Hub root command group `az iot device-update`.
-  To learn more about the service visit https://docs.microsoft.com/en-us/azure/iot-hub-device-update/.
+  To learn more about the service visit https://learn.microsoft.com/en-us/azure/iot-hub-device-update/.
 
   - This command group is behind a feature flag environment variable. Set `IOT_CLI_ADU_ENABLED` to any value
     to activate the command group.
@@ -824,7 +824,7 @@ Other notable changes, which are not affected by API versions, include:
     * az iot device send-d2c-message
     * az iot device simulate
 
-For more information about IoT Hub support for AAD visit: https://docs.microsoft.com/en-us/azure/iot-hub/iot-hub-dev-guide-azure-ad-rbac
+For more information about IoT Hub support for AAD visit: https://learn.microsoft.com/en-us/azure/iot-hub/iot-hub-dev-guide-azure-ad-rbac
 
 **Azure Digital Twins updates**