+++ HISTORY.rst
@@ -9,7 +9,7 @@ Release History
 **General updates**
 
 * We have dropped support for Python 3.8
-* This extension now supports Python 3.12 as the CLI core is packaging newer releases with this python version, 
+* This extension now supports Python 3.12 as the CLI core is packaging newer releases with this python version,
 * **[Breaking Change]** Older versions of `uamqp` (below `1.6.6`) are not compatible with Python 3.12
   * If you update your packaged AZ CLI version (`2.66.0` or later) or otherwise change your environment's Python version to 3.12, you will need to update your `uamqp` dependency to `1.6.6` or above in order to use commands like `az iot hub monitor-events`
   * You can repair this dependency at command runtime by utilizing the `--repair` / `-r` argument in `az iot hub monitor-events`
@@ -19,6 +19,10 @@ Release History
 * **[Breaking Change]** Device c2d messages (`az iot device c2d-message`) have been updated to support the following service-side changes:
   * `ContentEncoding` system property is now `content-encoding`
   * `ContentType` system property is now `content-type`
+
+**IoT Hub updates**
+
+* Removal of `az iot hub certificate root-authority`. For any issues, email iot-ca-updates@microsoft.com.
 
 
 0.25.0